Enhancing test suite for retry schedules #197.
